still trying to reorganize figbuilder and dataprocessor utility functions
