+++ app/pages/character_listing.py
@@ -5,7 +5,7 @@ from datetime import datetime as dt
 import pandas as pd
 
 import app.es.es_read_router as esr
-import app.figdata_manager.data_processor as dp 
+import app.data_service.field_flattener as fflat 
 import app.figdata_manager.color_meta as cm
 import app.pages.components as cmp
 from app.show_metadata import ShowKey
@@ -127,8 +127,8 @@ def render_speaker_listing_dt(show_key: str):
 
     indexed_speakers_response = esr.fetch_indexed_speakers(ShowKey(show_key), extra_fields='topics_mbti')
     indexed_speakers = indexed_speakers_response['speakers']
-    indexed_speakers = dp.flatten_speaker_topics(indexed_speakers, 'mbti', limit_per_speaker=3) 
-    indexed_speakers = dp.flatten_and_refine_alt_names(indexed_speakers, limit_per_speaker=1) 
+    indexed_speakers = fflat.flatten_speaker_topics(indexed_speakers, 'mbti', limit_per_speaker=3) 
+    indexed_speakers = fflat.flatten_and_refine_alt_names(indexed_speakers, limit_per_speaker=1) 
     
     speakers_df = pd.DataFrame(indexed_speakers)
